refactor : rouletteSlice 컴포넌트 분리
