+++ frontend/src/features/roulette/components/RouletteSlice/RouletteSlice.styled.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const PlayerNameText = styled.text`
+  fill: ${({ theme }) => theme.color.point[100]};
+  font-size: 12px;
+  font-weight: bold;
+`;
