+++ frontend/src/features/roulette/components/RouletteWheel/RouletteWheel.styled.ts
@@ -38,9 +38,3 @@ export const Wrapper = styled.div<WrapperProps>`
     }
   }
 `;
-
-export const PlayerNameText = styled.text`
-  fill: ${({ theme }) => theme.color.point[100]};
-  font-size: 12px;
-  font-weight: bold;
-`;
