+++ frontend/src/features/roulette/components/RouletteWheel/RouletteWheel.tsx
@@ -1,11 +1,9 @@
 import { useTheme } from '@emotion/react';
-import { describeArc } from '../../utils/describeArc';
 import { Angle, PlayerProbability } from '@/types/roulette';
 import * as S from './RouletteWheel.styled';
-import { getCenterAngle } from '../../utils/getCenterAngle';
-import { getTextPosition } from '../../utils/getTextPosition';
 import { WHEEL_CONFIG } from '../../constants/config';
 import { convertProbabilitiesToAngles } from '../../utils/convertProbabilitiesToAngles';
+import RouletteSlice from '../RouletteSlice/RouletteSlice';
 
 type Props =
   | {
@@ -46,38 +44,3 @@ const RouletteWheel = ({ angles, playerProbabilities, isSpinning = false }: Prop
 };
 
 export default RouletteWheel;
-
-type RouletteSliceProps = {
-  player: Angle;
-  strokeColor: string;
-};
-
-const RouletteSlice = ({ player, strokeColor }: RouletteSliceProps) => {
-  const centerAngle = getCenterAngle(player.startAngle, player.endAngle);
-  const textPosition = getTextPosition(centerAngle);
-
-  return (
-    <g key={player.playerName}>
-      <path
-        d={describeArc({
-          cx: WHEEL_CONFIG.CENTER,
-          cy: WHEEL_CONFIG.CENTER,
-          r: WHEEL_CONFIG.RADIUS,
-          startAngle: player.startAngle,
-          endAngle: player.endAngle,
-        })}
-        fill={player.playerColor}
-        stroke={strokeColor}
-        strokeWidth={WHEEL_CONFIG.STROKE_WIDTH}
-      />
-      <S.PlayerNameText
-        x={textPosition.x}
-        y={textPosition.y}
-        textAnchor="middle"
-        dominantBaseline="middle"
-      >
-        {player.playerName}
-      </S.PlayerNameText>
-    </g>
-  );
-};
